fix: rename-component
