+++ src/app/layout.tsx
@@ -1,8 +1,8 @@
 import RootLayout from "@/components/layout/RootLayout";
 
 export const metadata = {
-    title: "It.dev",
-    description: "Itのポートフォリオサイト",
+	title: "It.dev",
+	description: "Itのポートフォリオサイト",
 };
 
 export default RootLayout;
